# Improve performance for neighbor stats