Currently on PROD, the round pages take ~2.5 seconds each to load. >2 seconds of this time is coming from the logic to calculate a user's neighbors for the stats bar.

Not all users proof in all rounds, so the size of the users table is always going to be bigger than the number of tallies in a round. So instead of generating tallies in a round for all users and trimming out those who aren't present, only get users with tallies and special-case when the current user isn't in the tally round. Performance is now based on how many users have proofread in a round rather than the number of entries in the user table.

I've simplified the function which makes it specific to this one use-case, but it's only used in one place so the extensibility isn't necessary.

Here are the timings of the SQL for the 5 round pages on PROD:
| Round        | Current   | PR        |
|--------------|-----------|-----------|
| P1           | 2.07 sec  | 1.12 sec  |
| P2           | 1.75 sec  | 0.13 sec  |
| P3           | 1.90 sec  | 0.02 sec  |
| F1           | 1.77 sec  | 0.10 sec  |
| F2           | 1.82 sec  | 0.00 sec  |

Testable in https://www.pgdp.org/~cpeel/c.branch/improve-neighbor-performance/

Results should be identical to what is live on TEST now, although the performance difference will be negligible there. Be sure to look at a round page for a user who hasn't proofread anything in that round and confirm user's 0-entry is present and last.