@chrismiceli 2023-12-27
This performance improvement will be amazing!